requirements.txt
flask
sqlalchemy
python-docx
pandas
openpyxl
requests